fix mail reward in monster collection

+++ nekoyume/Assets/_Scripts/UI/Module/MailReward.cs
@@ -11,18 +11,18 @@ namespace Nekoyume.UI.Module
     {
         public ItemBase ItemBase { get; }
         public FungibleAssetValue FavFungibleAssetValue { get; }
-        public int Count { get; }
+        public long Count { get; }
 
         public bool IsPurchased { get; }
 
-        public MailReward(ItemBase itemBase, int count, bool isPurchased = false)
+        public MailReward(ItemBase itemBase, long count, bool isPurchased = false)
         {
             ItemBase = itemBase;
             Count = count;
             IsPurchased = isPurchased;
         }
 
-        public MailReward(FungibleAssetValue fav, int count, bool isPurchased = false)
+        public MailReward(FungibleAssetValue fav, long count, bool isPurchased = false)
         {
             FavFungibleAssetValue = fav;
             Count = count;
